requirements.txt
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework_jwt==1.11.0
whitenoise==5.3.0
Pillow

